modif bomp
