+++ app/Models/Ticket/Ticket.php
@@ -47,7 +47,6 @@ use Illuminate\Support\Facades\Auth;
  * @property Collection|Thread[] $threads
  * @property Collection|TagList[] $tagLists
  * @property Collection|Comment[] $comments
- * @property Collection|History[] $historicals
  * @property Channel $channel
  * @property Order $order
  * @property User $user
@@ -109,10 +108,6 @@ class Ticket extends Model
     public function comments(): HasMany
     {
         return $this->hasMany(Comment::class)->orderBy('id', 'DESC');
-    }
-    public function histories(): HasMany
-    {
-        return $this->hasMany(History::class)->orderBy('id', 'DESC');
     }
 
     public function getShowRoute(): string
@@ -195,7 +190,7 @@ class Ticket extends Model
 
         if ($mode != 'user') {
             $columns[] = (new TableColumnBuilder())
-                ->setLabel(__('app.ticket.user_id'))
+                ->setLabel(__('app.ticket.owner'))
                 ->setType(ColumnTypeEnum::SELECT)
                 ->setOptions(User::getUsersNames())
                 ->setCallback(function (Ticket $ticket) {
@@ -344,35 +339,5 @@ class Ticket extends Model
             ->where('tickets.channel_id', $channel->id)
             ->where('orders.channel_order_number', $order_number)
             ->first();
-    }
-
-    protected $historisable = [
-        'state',
-        'user_id',
-        'deadline',
-        'direct_customer_email',
-        'customer_issue',
-        'delivery_date'
-    ];
-
-    protected function addHistory(string $column, mixed $value): void
-    {
-        $historical = new History();
-        $historical->user_id = auth()->id();
-        $historical->type = $column;
-        $historical->modification = $value;
-        $historical->ticket_id = $this->id;
-        $historical->save();
-    }
-
-    public function save(array $options = [])
-    {
-        foreach($this->historisable as $column) {
-            if($this->isDirty($column)) {
-                $this->addHistory($column, $this->$column);
-            }
-        }
-
-        return parent::save($options);
     }
 }
